src/ml/data/OpenAlpha_Evolve/requirements.txt
# 核心套件
litellm
python-dotenv>=1.0.0
PyYAML==6.0
tiktoken
openai
anthropic
google-generativeai
gradio
docker
Jinja2
pydantic
watchdog
termcolor
requests
typing-extensions

# LLM 模型與框架
# transformers
# torch
# accelerate
# bitsandbytes # 若需要量化可使用
# sentencepiece

# 演化演算法 / 強化學習
# deap
# pygad
# ray[rllib]
# stable-baselines3

# 評估與沙盒
pytest
# hypothesis
# e2b-sdk  # E2B 沙盒

# 資料儲存
# psycopg2-binary  # PostgreSQL
# pymongo  # MongoDB
# dvc

# 工具
# jinja2

# API/Web（如需對外服務）
# fastapi
# uvicorn

# 其他可能用途或特定代理需求（例）
